Split vision runner and parameterize drift threshold
